Update block number, test and artifacts.
