--- tasks/20230519-l2-gauge-checkpointer/input.ts
@@ -1,14 +1,0 @@
-import { Task, TaskMode } from '@src';
-
-export type L2GaugeCheckpointerDeployment = {
-  GaugeController: string;
-  AuthorizerAdaptorEntrypoint: string;
-};
-
-const GaugeController = new Task('20220325-gauge-controller', TaskMode.READ_ONLY);
-const AuthorizerAdaptorEntrypoint = new Task('20221124-authorizer-adaptor-entrypoint', TaskMode.READ_ONLY);
-
-export default {
-  GaugeController,
-  AuthorizerAdaptorEntrypoint,
-};
