--- tasks/20230519-l2-gauge-checkpointer/readme.md
@@ -1,9 +1,0 @@
-# 2023-05-19 - L2 Gauge Checkpointer
-
-Deployment of the `L2GaugeCheckpointer` contract. It automates the process of performing checkpoints to stakeless root gauges.
-
-## Useful Files
-
-- [Ethereum mainnet addresses](./output/mainnet.json)
-- [Goerli testnet addresses](./output/goerli.json)
-- [`L2GaugeCheckpointer` artifact](./artifact/L2GaugeCheckpointer.json)
