+++ tasks/20230527-l2-gauge-checkpointer/readme.md
@@ -1,0 +1,7 @@
+# 2023-05-27 - L2 Gauge Checkpointer
+
+Deployment of the `L2GaugeCheckpointer` contract. It automates the process of performing checkpoints to stakeless root gauges.
+
+## Useful Files
+
+- [`L2GaugeCheckpointer` artifact](./artifact/L2GaugeCheckpointer.json)
